chore: move command_helper into util
Also remove some unused includes.
